# webapp/api-spec.yaml
openapi: 3.0.0
info:
  title: Dark Eccho Actions
  version: 1.1.0
paths:
  /text:
    get:
      operationId: getText
      summary: Récupérer un fichier en fonction du chemin variable
      parameters:
        - in: query
          name: path
          schema:
            type: string
          required: true
          description: Chemin vers le fichier (variable)
      responses:
        '200':
          description: Succès
          content:
            text/plain:
              example: |
                Contenu du fichier demandé.
        '404':
          description: Fichier non trouvé

  /query:
    get:
      operationId: getQuery
      summary: Récupère un élément de DynamoDB par une clé codée en base64
      parameters:
        - name: key
          in: query
          required: true
          schema:
            type: string
          description: Clé codée en base64
        - name: tableName
          in: query
          required: true
          schema:
            type: string
          description: Nom de la table DynamoDB
      responses:
        '200':
          description: Succès
        '409':
          description: Conflit, échec de la condition
        '500':
          description: Erreur serveur

  /scan:
    post:
      operationId: postScan
      summary: Scan une table dynamoDb
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                input:
                  type: object
                  properties:
                    AnyValue: 
                      description: Can be any value - string, number, boolean, array or object.
                  description: Objet DynamoDB à insérer ou mettre à jour
                tableName:
                  type: string
                  description: Nom de la table DynamoDB
      responses:
        '200':
          description: Élément trouvé
        '500':
          description: Erreur serveur

  /item:
    get:
      operationId: getItem
      summary: Récupère un élément de DynamoDB par son ID
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
          description: id de la clé de partition
        - name: tableName
          in: query
          required: true
          schema:
            type: string
          description: Nom de la table DynamoDB
        - name: pk
          in: query
          required: true
          schema:
            type: string
          description: clé de partition
      responses:
        '200':
          description: Succès
        '409':
          description: Conflit, échec de la condition
        '500':
          description: Erreur serveur
    post:
      operationId: postItem
      summary: Ajoute ou met à jour un élément dans DynamoDB
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                item:
                  type: object
                  properties:
                    AnyValue: 
                      description: Can be any value - string, number, boolean, array or object.
                  description: Objet DynamoDB à insérer ou mettre à jour
                tableName:
                  type: string
                  description: Nom de la table DynamoDB
                pk:
                  type: string
                  description: Clé de partition
      responses:
        '201':
          description: Élément créé ou mis à jour
        '409':
          description: Conflit, échec de la condition
        '500':
          description: Erreur serveur

servers:
  - url: 'https://actions.darkeccho.com'